# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ellowring API
  description: |
    Ellowring Software Solutions — Enterprise SaaS REST API (Phase 5).
    Learn. Prepare. Build. Get Hired.
  version: 1.0.0
  contact:
    name: Ellowring Engineering
    email: engineering@ellowring.com
servers:
  - url: http://localhost:4000/api
    description: Local
  - url: https://api-staging.ellowring.com/api
    description: Staging
  - url: https://api.ellowring.com/api
    description: Production
...
paths:
...
  /v1/courses/{id}/enroll:
    post:
      tags:
        - Student
      summary: Enroll in published course (shared).
      operationId: post__courses__id__enroll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/colleges/me:
    get:
      tags:
        - College
      summary: College profile for logged-in college user.
      operationId: get__colleges_me
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: city
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    patch:
      tags:
        - College
      summary: Update college profile fields.
      operationId: patch__colleges_me
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/colleges/me/applications:
    get:
      tags:
        - College
      summary: Incoming college applications.
      operationId: get__colleges_me_applications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: city
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/colleges/me/applications/{id}:
    patch:
      tags:
        - College
      summary: Update application status.
      operationId: patch__colleges_me_applications__id_
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/colleges/me/departments:
    get:
      tags:
        - College
      summary: List departments.
      operationId: get__colleges_me_departments
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: city
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    post:
      tags:
        - College
      summary: Create department.
      operationId: post__colleges_me_departments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/colleges/me/courses:
    get:
      tags:
        - College
      summary: List college courses/programmes.
      operationId: get__colleges_me_courses
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: city
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    post:
      tags:
        - College
      summary: Create college course.
      operationId: post__colleges_me_courses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/colleges:
    get:
      tags:
        - College
      summary: Public college catalogue.
      operationId: get__colleges
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: city
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/colleges/{id}:
    get:
      tags:
        - College
      summary: Public college detail.
      operationId: get__colleges__id_
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: city
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/courses:
    get:
      tags:
        - "Catalogue & Platform"
      summary: List published courses.
      operationId: get__courses
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: examType
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/courses/{slug}:
    get:
      tags:
        - "Catalogue & Platform"
      summary: Course detail by slug.
      operationId: get__courses__slug_
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: "string?"
          schema:
            type: string
        - name: examType
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data: {}
        meta:
          type: object
          properties:
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            pagination:
              $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        hasPrev: { type: boolean }